+++ semcore/chart/src/Chart/Venn/VennIntersection.tsx
@@ -4,6 +4,7 @@ import React from 'react';
 import { Layer } from 'recharts';
 import propsForElement from '@semcore/utils/lib/propsForElement';
 import { IVennChildProps } from './interface/VennChart';
+import { colors } from '../../utils/colors';
 
 export default class VennIntersection extends React.PureComponent<IVennChildProps<SVGPathElement>> {
   static displayName = 'VennIntersection';
@@ -11,7 +12,7 @@ export default class VennIntersection extends React.PureComponent<IVennChildProp
     fillOpacity: 0,
     activeFillOpacity: 0.1,
     strokeWidth: 2,
-    stroke: '#ffffff',
+    stroke: colors['white-01'],
   };
 
   render() {