+++ semcore/checkbox/src/Checkbox.js
@@ -10,12 +10,6 @@ import getInputProps, { inputProps } from '@semcore/utils/lib/inputProps';
 
 import style from './style/checkbox.shadow.css';
 
-const SIZE_TEXT_MAP = {
-  xl: 300,
-  l: 200,
-  m: 100,
-};
-
 class Checkbox extends Component {
   static displayName = 'Checkbox';
   static style = style;
@@ -28,7 +22,7 @@ class Checkbox extends Component {
   getTextProps() {
     const { size, disabled } = this.asProps;
     return {
-      size: size ? SIZE_TEXT_MAP[size] : size,
+      size,
       disabled,
     };
   }