+++ semcore/d3-chart/src/HorizontalBar.js
@@ -13,7 +13,6 @@ class HorizontalBarRoot extends Component {
   static style = style;
 
   static defaultProps = {
-    color: '#50aef4',
     offset: [0, 0],
     duration: 500,
     r: 2,