+++ website/docs/data-display/donut-chart/examples/donut.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Donut, Plot, Tooltip } from '@semcore/d3-chart';
+import { Donut, Plot, Tooltip, colors } from '@semcore/d3-chart';
 import { Flex } from '@semcore/flex-box';
 import { Text } from '@semcore/typography';
 
@@ -8,8 +8,8 @@ export default () => {
     <Plot width={300} height={300} data={data}>
       <Donut innerRadius={100}>
         <Donut.Pie dataKey="a" name="Pie 1" />
-        <Donut.Pie dataKey="b" color="#890C85" name="Pie 2" />
-        <Donut.Pie dataKey="c" color="#3AB011" name="Pie 3" />
+        <Donut.Pie dataKey="b" color={colors['green-02']} name="Pie 2" />
+        <Donut.Pie dataKey="c" color={colors['violet-04']} name="Pie 3" />
         <Donut.Label>Example</Donut.Label>
         <Tooltip>
           {({ dataKey, name }) => {