+++ website/docs/data-display/donut-chart/examples/semi-donut-with-one-data.js
@@ -7,9 +7,9 @@ export default () => {
     <Plot width={300} height={150} data={data}>
       <Donut halfsize innerRadius={100}>
         <Donut.Pie dataKey="speed" />
-        <Donut.Pie dataKey="other" color="#E4ECF1" />
+        <Donut.Pie dataKey="other" color="#C4C7CF" />
         <Donut.Label>
-          <Text tag="tspan" x="0" dy="-1.2em" fill="#757575" size={400}>
+          <Text tag="tspan" x="0" dy="-1.2em" fill="#6C6E79" size={400}>
             Keyword volume
           </Text>
         </Donut.Label>