+++ semcore/counter/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0] - 2021-12-08
+
+### Added
+
+- Added warning and danger themes
+
 ## [1.2.1] - 2021-8-26
 
 ### Changed